Move ConfigFile() and ConfigDir() from package main to packer
