+++ packer/config_file_unix.go
@@ -1,6 +1,6 @@
 // +build darwin freebsd linux netbsd openbsd
 
-package main
+package packer
 
 import (
 	"bytes"
