+++ packer/config_file_windows.go
@@ -1,6 +1,6 @@
 // +build windows
 
-package main
+package packer
 
 import (
 	"path/filepath"
